requirements.txt
flask==2.0.1
pydantic==1.8.2
pillow==9.0.0
psd-tools==1.9.24
openai==0.27.0
requests==2.28.1
python-dotenv==0.19.1